Move generic tests into namespace
