+++ apps/dav/tests/unit/CalDAV/Activity/Setting/GenericTest.php
@@ -19,7 +19,7 @@
  *
  */
 
-namespace OCA\DAV\Tests\unit\CalDAV\Activity;
+namespace OCA\DAV\Tests\unit\CalDAV\Activity\Setting;
 
 use OCA\DAV\CalDAV\Activity\Setting\Calendar;
 use OCA\DAV\CalDAV\Activity\Setting\Event;
@@ -27,7 +27,7 @@ use OCA\DAV\CalDAV\Activity\Setting\Todo;
 use OCP\Activity\ISetting;
 use Test\TestCase;
 
-class GenericSettingTest extends TestCase {
+class GenericTest extends TestCase {
 
 	public function dataSettings() {
 		return [
